+++ plugin/touchify-color-options/ColorOptionsDocker.py
@@ -2,10 +2,10 @@
 from krita import *
 from PyQt5.QtCore import *
 
+from touchify.src.components.touchify.special.CanvasColorPicker import CanvasColorPicker
 from touchify.src.settings import TouchifyConfig
 from touchify.src.variables import *
 
-from touchify.src.components.pyqt.widgets.ColorFramedButton import ColorFramedButton
 from touchify.src.resources import ResourceManager
 from touchify.src.helpers import TouchifyHelpers
 
@@ -34,9 +34,8 @@ class ColorSourceToggle(QWidget):
         self.gridLayout.setSpacing(0)
         self.gridLayout.setContentsMargins(0,0,0,0)
 
-        self.setFgBtn = ColorFramedButton(self)
+        self.setFgBtn = CanvasColorPicker(self, CanvasColorPicker.Mode.Foreground)
         self.setFgBtn.setFocusPolicy(Qt.FocusPolicy.NoFocus)
-        self.setFgBtn.clicked.connect(self.setForegroundColor)
         self.gridLayout.addWidget(self.setFgBtn)
 
         self.toggleBtn = QPushButton(self)
@@ -45,9 +44,8 @@ class ColorSourceToggle(QWidget):
         self.toggleBtn.clicked.connect(self.toggleColors)
         self.gridLayout.addWidget(self.toggleBtn)
 
-        self.setBgBtn = ColorFramedButton(self)
+        self.setBgBtn = CanvasColorPicker(self, CanvasColorPicker.Mode.Background)
         self.setBgBtn.setFocusPolicy(Qt.FocusPolicy.NoFocus)
-        self.setBgBtn.clicked.connect(self.setBackgroundColor)
         self.gridLayout.addWidget(self.setBgBtn)
 
         self.resetBtn = QPushButton(self)
@@ -71,28 +69,22 @@ class ColorSourceToggle(QWidget):
 
         self.setFgBtn.setFixedHeight(cubeSize)
         self.setBgBtn.setFixedHeight(cubeSize)
-        
-
 
     def setup(self, instance: "TouchifyWindow"):
         self.sourceWindow: Window = instance.windowSource
         parentExtension = TouchifyHelpers.getExtension()
         if parentExtension:
-            parentExtension.intervalTimerTicked.connect(self.updateColors)
-
-    def setForegroundColor(self):
-        Krita.instance().action("chooseForegroundColor").trigger()
-
-    def setBackgroundColor(self):
-        Krita.instance().action("chooseBackgroundColor").trigger()
+            parentExtension.intervalTimerTicked.connect(self.intervalTimerTicked)
 
     def toggleColors(self):
         Krita.instance().action("toggle_fg_bg").trigger()
-        self.updateColors()
+        self.setFgBtn.updateColors()
+        self.setBgBtn.updateColors()
 
     def resetColors(self):
         Krita.instance().action("reset_fg_bg").trigger()
-        self.updateColors()
+        self.setFgBtn.updateColors()
+        self.setBgBtn.updateColors()
 
     def showEvent(self, event):
         super().showEvent(event)
@@ -100,25 +92,12 @@ class ColorSourceToggle(QWidget):
     def closeEvent(self, event):
         super().closeEvent(event)
 
-    def krita_to_qcolor(self, source: ManagedColor):
-        if self.canvas == None or source == None: return QColor()
-        return source.colorForCanvas(self.canvas)
-
-    def updateColors(self):
-        activeWin = self.sourceWindow
-        if activeWin == None: return
-        activeView = activeWin.activeView()
-        if activeView == None: return
-        
-        fg_color = self.krita_to_qcolor(activeView.foregroundColor())
-        bg_color = self.krita_to_qcolor(activeView.backgroundColor())
-        
-        self.setFgBtn.setColor(fg_color)
-        self.setBgBtn.setColor(bg_color)
+    def intervalTimerTicked(self):
+        pass
 
     def onCanvasChanged(self, canvas: Canvas):
-        self.canvas = canvas
-        self.updateColors()
+        self.setFgBtn.updateColors()
+        self.setBgBtn.updateColors()
 
 class ColorOptionsDocker(DockWidget):